Create sensor
