+++ homeassistant/components/proximity/const.py
@@ -12,6 +12,7 @@ DOMAIN = "proximity"
 ATTR_DIR_OF_TRAVEL = "dir_of_travel"
 ATTR_DIST_FROM = "dist_to_zone"
 ATTR_NEAREST = "nearest"
+ATTR_DEVICES = "devices"
 
 CONF_IGNORED_ZONES = "ignored_zones"
 CONF_TOLERANCE = "tolerance"
@@ -22,6 +23,8 @@ DEFAULT_NEAREST = "not set"
 DEFAULT_PROXIMITY_ZONE = "home"
 DEFAULT_TOLERANCE = 1
 DOMAIN = "proximity"
+
+ICON = "hass:apple-safari"
 
 UNITS = [
     LENGTH_METERS,
